stl3d-gui/requirements.txt
numpy>=1.26.0; python_version >= "3.9"
numpy>=1.21.6; python_version < "3.9"
pillow>=8.0.0
numpy-stl>=2.16.0
opencv-python>=4.5.0
matplotlib>=3.4.0
scipy>=1.7.0
scikit-image>=0.18.0
tqdm>=4.60.0
trimesh>=3.9.0
pyglet>=1.5.0